+++ release-controller/pytest.py
@@ -2,8 +2,12 @@ import os
 
 if __name__ == "__main__":
     import pytest
+    import sys
 
     dir_path = os.path.dirname(__file__)
+    sys.path.append(dir_path)
+    sys.path.append(os.path.join(dir_path, "tests"))
+
     if hasattr(pytest, "main"):
         raise SystemExit(
             pytest.main(
